+++ node/graphql/index.ts
@@ -6,7 +6,7 @@ const getAppId = (): string => {
   return process.env.VTEX_APP_ID ?? ''
 }
 
-const SCHEMA_VERSION = 'v6.7'
+const SCHEMA_VERSION = 'v6.8'
 
 const routes = {
   baseUrl: (account: string) =>
@@ -20,11 +20,11 @@ const routes = {
   listCarts: (account: string, email: string) =>
     `${routes.cartEntity(
       account
-    )}/search?email=${email}&_schema=${SCHEMA_VERSION}&_fields=id,email,cartName,items,creationDate,subtotal,discounts,shipping,total,customData,address&_sort=creationDate DESC`,
+    )}/search?email=${email}&_schema=${SCHEMA_VERSION}&_fields=id,email,cartName,status,description,items,creationDate,subtotal,discounts,shipping,total,customData,address&_sort=creationDate DESC`,
   getCart: (account: string, id: string) =>
     `${routes.cartEntity(
       account
-    )}/documents/${id}?_fields=id,email,cartName,items,creationDate,subtotal,discounts,shipping,total,customData,address`,
+    )}/documents/${id}?_fields=id,email,cartName,status,description,items,creationDate,subtotal,discounts,shipping,total,customData,address`,
 
   saveSchema: (account: string) =>
     `${routes.cartEntity(account)}/schemas/${SCHEMA_VERSION}`,
@@ -48,6 +48,14 @@ const schema = {
       type: 'string',
       title: 'Cart Name',
     },
+    status: {
+      type: ['null', 'string'],
+      title: 'Status',
+    },
+    description: {
+      type: ['null', 'string'],
+      title: 'Description',
+    },
     items: {
       type: 'array',
       title: 'Cart',
@@ -61,7 +69,7 @@ const schema = {
       title: 'Cart Life Span',
     },
     subtotal: {
-      type: 'float',
+      type: 'number',
       title: 'Subtotal',
     },
     discounts: {
@@ -77,11 +85,11 @@ const schema = {
       title: 'Custom Data',
     },
     total: {
-      type: ['number', 'float'],
+      type: 'number',
       title: 'Total',
     },
   },
-  'v-indexed': ['email', 'creationDate', 'cartLifeSpan', 'cartName'],
+  'v-indexed': ['email', 'creationDate', 'cartLifeSpan', 'cartName', 'status'],
   'v-default-fields': ['email', 'cart', 'creationDate', 'cartLifeSpan'],
   'v-cache': false,
 }