+++ CHANGELOG.md
@@ -7,9 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Description and Status to the quote
+
+### Updated
+
+- Changing schema subtotal data type from `float` to `number`
+
 ## [1.2.4] - 2021-01-12
 
 ### Changed
+
 - Make cart subtotal data type from `integer` to `float`
 - Updated graphql schema version